Added query options type
